Restrict simuls to one-board variants and harden simul state

+++ server/game.py
@@ -573,7 +573,12 @@ class Game:
 
         asyncio.create_task(self.app_state.remove_from_cache(self), name="game-remove-%s" % self.id)
 
-        if self.board.ply < 3 and (self.app_state.db is not None) and (self.tournamentId is None):
+        if (
+            self.board.ply < 3
+            and (self.app_state.db is not None)
+            and (self.tournamentId is None)
+            and (self.simulId is None)
+        ):
             result = await self.app_state.db.game.delete_one({"_id": self.id})
             log.debug(
                 "Removed too short game %s from db. Deleted %s game.",
